requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
email-validator==2.0.0
python-dotenv==1.0.0
pdfminer.six==20221105
python-docx==0.8.11
Werkzeug==2.2.3
Flask-Bcrypt==1.0.1
pandas==1.5.3
gunicorn